chore: separate files for types
