+++ packages/types/src/context.d.ts
@@ -1,0 +1,5 @@
+import type { AttributeKey, AttributeValue } from "./attribute";
+
+export interface Context {
+  [key: AttributeKey]: AttributeValue;
+}
