+++ packages/types/src/index.d.ts
@@ -1,120 +1,10 @@
-export type AttributeKey = string;
+import { Attribute, AttributeKey, AttributeValue } from "./attribute";
+import { Condition } from "./condition";
+import { Context } from "./context";
 
-export interface AttributeObjectValue {
-  [key: AttributeKey]: AttributeValue;
-}
-
-export type AttributeValue =
-  | string
-  | number
-  | boolean
-  | Date
-  | null
-  | undefined
-  | string[]
-  | AttributeObjectValue;
-
-export interface Context {
-  [key: AttributeKey]: AttributeValue;
-}
-
-export type AttributeType =
-  | "boolean"
-  | "string"
-  | "integer"
-  | "double"
-  | "date"
-  | "semver"
-  | "object"
-  | "array";
-
-export interface Attribute {
-  archived?: boolean; // only available in YAML files
-  key?: AttributeKey; // needed for supporting v1 datafile generation
-  type: AttributeType;
-  description?: string; // only available in YAML files
-  properties?: {
-    [key: AttributeKey]: {
-      type:
-        | "boolean"
-        | "string"
-        | "integer"
-        | "double"
-        | "date"
-        | "semver"
-        // | "object" // NOTE: avoid nesting for now
-        | "array";
-      description?: string;
-    };
-  };
-}
-
-export type Operator =
-  | "equals"
-  | "notEquals"
-  | "exists"
-  | "notExists"
-
-  // numeric
-  | "greaterThan"
-  | "greaterThanOrEquals"
-  | "lessThan"
-  | "lessThanOrEquals"
-
-  // string
-  | "contains"
-  | "notContains"
-  | "startsWith"
-  | "endsWith"
-
-  // semver (string)
-  | "semverEquals"
-  | "semverNotEquals"
-  | "semverGreaterThan"
-  | "semverGreaterThanOrEquals"
-  | "semverLessThan"
-  | "semverLessThanOrEquals"
-
-  // date comparisons
-  | "before"
-  | "after"
-
-  // array of strings
-  | "includes"
-  | "notIncludes"
-
-  // regex
-  | "matches"
-  | "notMatches"
-
-  // array of strings
-  | "in"
-  | "notIn";
-
-export type ConditionValue = string | number | boolean | Date | null | undefined | string[];
-
-export interface PlainCondition {
-  attribute: AttributeKey;
-  operator: Operator;
-  value?: ConditionValue; // for all operators, except for "exists" and "notExists"
-  regexFlags?: string; // for regex operators only (matches, notMatches)
-}
-
-export interface AndCondition {
-  and: Condition[];
-}
-
-export interface OrCondition {
-  or: Condition[];
-}
-
-export interface NotCondition {
-  not: Condition[];
-}
-
-export type AndOrNotCondition = AndCondition | OrCondition | NotCondition;
-
-export type Condition = PlainCondition | AndOrNotCondition | string;
+export * from "./attribute";
+export * from "./context";
+export * from "./condition";
 
 export type SegmentKey = string;
